feat: add route /c-by-id/<projectId>/preview and redirect /projects-by-id/<projectId>/preview
